+++ server/queues/processors/NotificationsProcessor.ts
@@ -1,29 +1,23 @@
 import { subHours } from "date-fns";
-import { uniqBy } from "lodash";
-import { Node } from "prosemirror-model";
 import { Op } from "sequelize";
+import { Minute } from "@shared/utils/time";
 import subscriptionCreator from "@server/commands/subscriptionCreator";
 import { sequelize } from "@server/database/sequelize";
-import { schema } from "@server/editor";
 import CollectionNotificationEmail from "@server/emails/templates/CollectionNotificationEmail";
-import CommentCreatedEmail from "@server/emails/templates/CommentCreatedEmail";
 import DocumentNotificationEmail from "@server/emails/templates/DocumentNotificationEmail";
 import env from "@server/env";
 import Logger from "@server/logging/Logger";
 import {
-  View,
   Document,
   Team,
   Collection,
-  User,
-  NotificationSetting,
-  Subscription,
   Notification,
   Revision,
-  Comment,
+  User,
+  View,
 } from "@server/models";
 import DocumentHelper from "@server/models/helpers/DocumentHelper";
-import ProsemirrorHelper from "@server/models/helpers/ProsemirrorHelper";
+import NotificationHelper from "@server/models/helpers/NotificationHelper";
 import {
   CollectionEvent,
   RevisionEvent,
@@ -31,6 +25,7 @@ import {
   DocumentEvent,
   CommentEvent,
 } from "@server/types";
+import CommentCreatedNotificationTask from "../tasks/CommentCreatedNotificationTask";
 import BaseProcessor from "./BaseProcessor";
 
 export default class NotificationsProcessor extends BaseProcessor {
@@ -56,84 +51,9 @@ export default class NotificationsProcessor extends BaseProcessor {
   }
 
   async commentCreated(event: CommentEvent) {
-    const [document, comment, team] = await Promise.all([
-      Document.scope("withCollection").findOne({
-        where: {
-          id: event.documentId,
-        },
-      }),
-      Comment.findByPk(event.modelId),
-      Team.findByPk(event.teamId),
-    ]);
-    if (!document || !comment || !team) {
-      return;
-    }
-
-    // Commenting on a doc automatically creates a subscription to the doc
-    // if they haven't previously had one.
-    await sequelize.transaction(async (transaction) => {
-      await subscriptionCreator({
-        user: comment.createdBy,
-        documentId: document.id,
-        event: "documents.update",
-        resubscribe: false,
-        transaction,
-        ip: event.ip,
-      });
+    await CommentCreatedNotificationTask.schedule(event, {
+      delay: Minute,
     });
-
-    const recipients = await this.getCommentNotificationRecipients(
-      document,
-      comment,
-      comment.createdById
-    );
-    if (!recipients.length) {
-      return;
-    }
-
-    let content = ProsemirrorHelper.toHTML(
-      Node.fromJSON(schema, comment.data),
-      {
-        centered: false,
-      }
-    );
-    if (!content) {
-      return;
-    }
-
-    content = await DocumentHelper.attachmentsToSignedUrls(
-      content,
-      event.teamId,
-      86400 * 4
-    );
-
-    for (const recipient of recipients) {
-      const notify = await this.shouldNotify(document, recipient.user);
-
-      if (notify) {
-        const notification = await Notification.create({
-          event: event.name,
-          userId: recipient.user.id,
-          actorId: comment.createdById,
-          teamId: team.id,
-          documentId: document.id,
-        });
-        await CommentCreatedEmail.schedule(
-          {
-            to: recipient.user.email,
-            documentId: document.id,
-            teamUrl: team.url,
-            isReply: !!comment.parentCommentId,
-            actorName: comment.createdBy.name,
-            commentId: comment.id,
-            content,
-            collectionName: document.collection?.name,
-            unsubscribeUrl: recipient.unsubscribeUrl,
-          },
-          { notificationId: notification.id }
-        );
-      }
-    }
   }
 
   async documentPublished(event: DocumentEvent) {
@@ -158,7 +78,7 @@ export default class NotificationsProcessor extends BaseProcessor {
 
     await this.createDocumentSubscriptions(document, event);
 
-    const recipients = await this.getDocumentNotificationRecipients(
+    const recipients = await NotificationHelper.getDocumentNotificationRecipients(
       document,
       "documents.publish",
       document.lastModifiedById
@@ -205,7 +125,7 @@ export default class NotificationsProcessor extends BaseProcessor {
 
     await this.createDocumentSubscriptions(document, event);
 
-    const recipients = await this.getDocumentNotificationRecipients(
+    const recipients = await NotificationHelper.getDocumentNotificationRecipients(
       document,
       "documents.update",
       document.lastModifiedById
@@ -263,7 +183,7 @@ export default class NotificationsProcessor extends BaseProcessor {
       return;
     }
 
-    const recipients = await this.getCollectionNotificationRecipients(
+    const recipients = await NotificationHelper.getCollectionNotificationRecipients(
       collection,
       event.name
     );
@@ -283,159 +203,10 @@ export default class NotificationsProcessor extends BaseProcessor {
     }
   }
 
-  /**
-   * Create any new subscriptions that might be missing for collaborators in the
-   * document on publish and revision creation. This does mean that there is a
-   * short period of time where the user is not subscribed after editing until a
-   * revision is created.
-   *
-   * @param document The document to create subscriptions for
-   * @param event The event that triggered the subscription creation
-   */
-  private createDocumentSubscriptions = async (
-    document: Document,
-    event: DocumentEvent | RevisionEvent
-  ): Promise<void> => {
-    await sequelize.transaction(async (transaction) => {
-      const users = await document.collaborators({ transaction });
-
-      for (const user of users) {
-        await subscriptionCreator({
-          user,
-          documentId: document.id,
-          event: "documents.update",
-          resubscribe: false,
-          transaction,
-          ip: event.ip,
-        });
-      }
-    });
-  };
-
-  /**
-   * Get the recipients of a notification for a collection event.
-   *
-   * @param collection The collection to get recipients for
-   * @param eventName The event name
-   * @returns A list of recipients
-   */
-  private getCollectionNotificationRecipients = async (
-    collection: Collection,
-    eventName: string
-  ): Promise<NotificationSetting[]> => {
-    // First find all the users that have notifications enabled for this event
-    // type at all and aren't the one that performed the action.
-    const recipients = await NotificationSetting.scope("withUser").findAll({
-      where: {
-        userId: {
-          [Op.ne]: collection.createdById,
-        },
-        teamId: collection.teamId,
-        event: eventName,
-      },
-    });
-
-    // Ensure we only have one recipient per user as a safety measure
-    return uniqBy(recipients, "userId");
-  };
-
-  private getCommentNotificationRecipients = async (
-    document: Document,
-    comment: Comment,
-    actorId: string
-  ): Promise<NotificationSetting[]> => {
-    const recipients = await this.getDocumentNotificationRecipients(
-      document,
-      "documents.update",
-      actorId
-    );
-
-    if (recipients.length > 0 && comment.parentCommentId) {
-      const contextComments = await Comment.findAll({
-        attributes: ["createdById"],
-        where: {
-          [Op.or]: [
-            { id: comment.parentCommentId },
-            { parentCommentId: comment.parentCommentId },
-          ],
-        },
-      });
-
-      const userIdsInThread = contextComments.map((c) => c.createdById);
-      return recipients.filter((r) => userIdsInThread.includes(r.userId));
-    }
-
-    return recipients;
-  };
-
-  /**
-   * Get the recipients of a notification for a document event.
-   *
-   * @param document The document to get recipients for
-   * @param eventName The event name
-   * @param actorId The id of the user that performed the action
-   * @returns A list of recipients
-   */
-  private getDocumentNotificationRecipients = async (
-    document: Document,
-    eventName: string,
-    actorId: string
-  ): Promise<NotificationSetting[]> => {
-    // First find all the users that have notifications enabled for this event
-    // type at all and aren't the one that performed the action.
-    let recipients = await NotificationSetting.scope("withUser").findAll({
-      where: {
-        userId: {
-          [Op.ne]: actorId,
-        },
-        teamId: document.teamId,
-        event: eventName,
-      },
-    });
-
-    // If the event is a revision creation we can filter further to only those
-    // that have a subscription to the document…
-    if (eventName === "documents.update") {
-      const subscriptions = await Subscription.findAll({
-        attributes: ["userId"],
-        where: {
-          userId: recipients.map((recipient) => recipient.user.id),
-          documentId: document.id,
-          event: eventName,
-        },
-      });
-
-      const subscribedUserIds = subscriptions.map(
-        (subscription) => subscription.userId
-      );
-
-      recipients = recipients.filter((recipient) =>
-        subscribedUserIds.includes(recipient.user.id)
-      );
-    }
-
-    // Ensure we only have one recipient per user as a safety measure
-    return uniqBy(recipients, "userId");
-  };
-
   private shouldNotify = async (
     document: Document,
     user: User
   ): Promise<boolean> => {
-    // Suppress notifications for suspended and users with no email address
-    if (user.isSuspended || !user.email) {
-      return false;
-    }
-
-    // Check the recipient has access to the collection this document is in. Just
-    // because they are subscribed doesn't meant they still have access to read
-    // the document.
-    const collectionIds = await user.collectionIds();
-
-    if (!collectionIds.includes(document.collectionId)) {
-      return false;
-    }
-
     // Deliver only a single notification in a 12 hour window
     const notification = await Notification.findOne({
       order: [["createdAt", "DESC"]],
@@ -485,5 +256,34 @@ export default class NotificationsProcessor extends BaseProcessor {
     }
 
     return true;
+  };
+
+  /**
+   * Create any new subscriptions that might be missing for collaborators in the
+   * document on publish and revision creation. This does mean that there is a
+   * short period of time where the user is not subscribed after editing until a
+   * revision is created.
+   *
+   * @param document The document to create subscriptions for
+   * @param event The event that triggered the subscription creation
+   */
+  private createDocumentSubscriptions = async (
+    document: Document,
+    event: DocumentEvent | RevisionEvent
+  ): Promise<void> => {
+    await sequelize.transaction(async (transaction) => {
+      const users = await document.collaborators({ transaction });
+
+      for (const user of users) {
+        await subscriptionCreator({
+          user,
+          documentId: document.id,
+          event: "documents.update",
+          resubscribe: false,
+          transaction,
+          ip: event.ip,
+        });
+      }
+    });
   };
 }
